ARM: Rockchip: fix clang warning about passing 32-bit register operand
clangd warns that "Value size does not match register size specified by
the constraint and modifier".

Promote the subtraction result to unsigned long to fix this.

The 16 is likely unnecessary, but I am hesitant to remove it without testing.

Link: https://lore.kernel.org/r/20241106185724.2526-4-a.fatoum@pengutronix.de

+++ arch/arm/mach-rockchip/atf.c
@@ -59,7 +59,7 @@ static unsigned long load_elf64_image_phdr(const void *elf)
 										\
 	/* Setup an initial stack for EL2 */                                    \
 	asm volatile("msr sp_el2, %0" : :                                       \
-			"r" (SOC##_BAREBOX_LOAD_ADDRESS - 16) :                 \
+			"r" ((ulong)SOC##_BAREBOX_LOAD_ADDRESS - 16) :		\
 			"cc");                                                  \
 										\
 	bl31_entry(bl31, optee_load_address,                                    \
